Canonicalize any symlinks in debug file paths, not just the base
Without this change, gdb uses the path to the loaded object verbatim to look
for debug objects.  Presuming (I believe) that the runtime linker will have
canonicalized any paths it sees, and if that fails and the filename is a
symbolic link, resolves that link and tries the result.  I think presuming the
runtime linker has performed at least some canonicalizatio.

Under illumos this is not the case, rtld stores verbatim paths in the
linkmaps, and symbolic links may be in the directory portions (such as
'/lib/64/libc.so.1').

It is probably that it would actually be better to canonicalize these paths in
rtld, but since that would change the information given to audit and support
libraries (albeit likely beneficially), we're being conservative for the
moment.